fix EnginePicture.toImage; test picture.toImage().toByteData()
