+++ lib/web_ui/lib/src/engine/html_image_codec.dart
@@ -136,6 +136,15 @@ class HtmlImage implements ui.Image {
 
   @override
   Future<ByteData?> toByteData({ui.ImageByteFormat format = ui.ImageByteFormat.rawRgba}) {
+    if (format == ui.ImageByteFormat.rawRgba) {
+      final html.CanvasElement canvas = html.CanvasElement()
+        ..width = width
+        ..height = height;
+      final html.CanvasRenderingContext2D ctx = canvas.context2D;
+      ctx.drawImage(imgElement, 0, 0);
+      final html.ImageData imageData = ctx.getImageData(0, 0, width, height);
+      return Future.value(imageData.data.buffer.asByteData());
+    }
     if (imgElement.src?.startsWith('data:') == true) {
       final data = UriData.fromUri(Uri.parse(imgElement.src!));
       return Future.value(data.contentAsBytes().buffer.asByteData());
